src/utils/resiliency_metrics/requirements.txt
google-cloud-logging>=3.0.0
google-auth>=2.0.0
pandas>=1.3.0
tabulate>=0.8.0
python-dateutil>=2.8.0
